Modularise handlers.
